Explicitly truncate bswap operand to uint32_t
There are some places in git where a long is passed to htonl/ntohl. llvm
doesn't support matching operands of different bitwidths intentionally.
This patch fixes the build with llvm-gcc (and clang) on x86_64.


cregit: b073b7a990deb1cb3425db45642fa18c8b3cb65c http://github.com/git/git/commit/b073b7a990deb1cb3425db45642fa18c8b3cb65c

+++ compat/bswap.h
@@ -89,7 +89,7 @@ name|bswap32
 parameter_list|(
 name|x
 parameter_list|)
-value|({ \ 	uint32_t __res; \ 	if (__builtin_constant_p(x)) { \ 		__res = default_swab32(x); \ 	} else { \ 		__asm__("bswap %0" : "=r" (__res) : "0" (x)); \ 	} \ 	__res; })
+value|({ \ 	uint32_t __res; \ 	if (__builtin_constant_p(x)) { \ 		__res = default_swab32(x); \ 	} else { \ 		__asm__("bswap %0" : "=r" (__res) : "0" ((uint32_t)(x))); \ 	} \ 	__res; })
 end_define
 begin_elif
 elif|#
